+++ fluence-faas/src/raw_toml_config.rs
@@ -115,7 +115,7 @@ pub struct TomlFaaSModuleConfig {
     pub logger_enabled: Option<bool>,
     pub wasi: Option<TomlWASIConfig>,
     pub mounted_binaries: Option<toml::value::Table>,
-    pub logging_mask: Option<i64>,
+    pub logging_mask: Option<i32>,
 }
 
 impl TomlFaaSNamedModuleConfig {
@@ -182,7 +182,7 @@ pub fn from_toml_module_config(config: TomlFaaSModuleConfig) -> Result<FaaSModul
         logger_enabled: config.logger_enabled.unwrap_or(true),
         host_imports: host_cli_imports,
         wasi,
-        logging_mask: config.logging_mask.unwrap_or(i64::max_value()),
+        logging_mask: config.logging_mask.unwrap_or(i32::max_value()),
     })
 }
